[nrf fromlist] dfu: Allow to use imgtool-based headers
It is possible to add MCUboot header through --pad-header option.
In such cases, the FLASH_LOAD_OFFSET does not point to the beginning of
the slot, but to the beginning of the executable area, thus the check
for the active slot should use ranges instead of exact values.

Upstream PR #: 96925

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>