[nrf fromlist] soc: nordic: nrf54h: s2ram: Use ARM SCB save/restore funcs
This reduced the amount of duplicate code and unifies the code
with other platforms.

With this change fewer registers are stored and restored.
See also comment in scb.h for scb_context stating that
only essential registers are stored and restored.

No longer stored:
- ICSR
- SCR
- CFSR
- HFSR
- DFSR
- MMFAR
- BFAR
- AFSR

No longer used:
- SHPR[3..12]. This backup register was declared in the wrong way.
  In core_cm33.h and core_cm4.h this is declared as an array of
  12 uint8_t's. That is 3 uint32_t's.

Orignal SCB retention was added in
2055f7d5956aecf55c4babca22eadb23d59bc534.

Upstream PR #: 97073

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>