[nrf fromlist] modem: pipe: Don't return EPERM on closed pipe
When working on CMUX power saving, it is typical
that we end up closing the pipe before the last
RX_READY event is handled from workqueue, so we end up
receiving -EPERM which is not really a fatal error.

Pipes recover when they are re-opened. So drop this error
and return zero instead, like modem_pipe_open() and close() does.

Upstream PR #: 97362

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>