dts: drivers: nordic: Don't manage hsfll120 from drivers
The hsfll120 is requested automatically by hardware which needs it,
CLOCK_CONTROL can be used to set to lower the frequency it runs at,
but this is a quirk which may be handled by the application if
desired. Disable the global clocks and remove references to the
hsfll120 from devices in the devicetree, and update samples which
specifically test the control of the clocks using CLOCK_CONTROL
to enable them.

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>