[nrf fromtree] Bluetooth: Mesh: Filter duplicates in brg subnets list
This commit adds functionality to filter out duplicate entries in the
Bridged Subnets List message.

This is done by iterating through the part of the table we have already
processed, to check if this is the first time we see a given key pair or
not.

Signed-off-by: Ludvig Jordet <ludvig.jordet@nordicsemi.no>
(cherry picked from commit cc8ac6bdee0bd58248ab877cce69475fa52ebc42)
Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>