[nrf fromtree] Bluetooth: Host: Make use of common array helper macros
There's a bunch of convenience macros in sys/util.h that are intended to
help out with translating array indices and determining if an element is
part of an array or not. Use those instead of custom code.

Signed-off-by: Johan Hedberg <johan.hedberg@silabs.com>
(cherry picked from commit 0fd8af97fbeab36159c732ae09d33d987d635197)
Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>