[nrf noup] ironside: add initial IronSide SE support package
This patch was ported to the 3.2 branch from a commit in the hal_nordic
repository where the IronSide SE support code is located on the main
branch.
It places the IronSide support code inside zephyr instead of pulling it
in by updating the hal_nordic revision. It is done this way to prevent
other code in hal_nordic (mainly nrfx) from being updated as a
side effect, since we are planning updates to this code and it is known
that updating nrfx past this point will make it necessary to add a lot
of additional patches.

The patch was made using folllowing commands:

cd modules/hal/nordic

git format-patch -1 a68b40b0b364668230a850e70867428269c0b915 --stdout > \
ironside_initial_0.patch

cd ../../../zephyr

mkdir modules/hal_nordic/ironside/se/support_package

git am --directory=modules/hal_nordic/ironside/se/support_package \
../modules/hal/nordic/ironside_initial_0.patch

See the original commit message below.

---

Rev: 3ed8e149a79697a0e8e58f60170b7b4df30acd82

Adds a new set of source files/scripts used to interact with
IronSide SE. The plan is for this to replace much of the
implementation that currently lives in zephyr/soc/nordic/ironside once
it is fully integrated.

The initial package includes:

* Serialization code for most IronSide calls
* Boot report definition
* UICR definition
* Macros for encoding PERIPHCONF entries
* Addresses for certain IronSide SE defined structures in RAM
  (IPC buffers, boot reports, event reports)
* Macros definining the released IronSide SE versions, including
  a concept of "target" IronSide SE version which will enable
  compatibility with older IronSide SE versions.
* Python code for generating UICR and PERIPHCONF blobs
* Python code for validating a PERIPHCONF blob

For now, the code is developed and tested in a different repository and
will be copied into this repository on a regular basis. A git sha is
used to track the origin commit/tag in that repository, similarly to
other projects integrated in hal_nordic. The origin SHA is also
contained in ironside/se/git_sha.txt.

Signed-off-by: Jonathan Nilsen <jonathan.nilsen@nordicsemi.no>