[nrf fromtree] west: Sync picolibc with SDK 0.17.4
This updates the version of picolibc to match that used by SDK 0.17.4. In
the Zephyr picolibc fork, that's marked with the zephyr-sdk-0.17.4 tag now,
which is on the zephyr-sdk-0.17 branch.

Changes since the previous version which impact using the module:

 * machine/arm: Disable exception tables in ARM string asm code
 * cmake: Silence messages about core-isa.h
 * Correct return type of __non_atomic_*_ungetc functions
 * Delete obsoleted _syslist.h

Signed-off-by: Keith Packard <keithp@keithp.com>
(cherry picked from commit 046232a6497463162fb846da4ba889ad1feff6cf)