[nrf noup] drivers: can: mcan: implement CAN_MODE_ONE_SHOT
Implement CAN_MODE_ONE_SHOT in MCAN impementation. The implementation
contains a workaround for a bug in the MCAN IP which prevents an IRQ
from triggering.

This is a noup as the workaround is too complicated and the feature
is too niche to be accepted it upstream.

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>
Signed-off-by: Tomasz Moń <tomasz.mon@nordicsemi.no>
(cherry picked from commit 0a0d1c2ed284424594da6c8fa631c7be02cd2a8c)