[nrf noup] Kconfig: Add support for relocation pattern in nRF54H20
This update introduces a new configuration option to automatically calculate
the Load Memory Address (LMA) adjustment for firmware relocation.
If the 'fw-to-relocate' chosen node exists, the firmware will load
from the MRAM partition it points to and run from the specified TCM region.
Otherwise, it defaults to the standard code-partition calculation.

Signed-off-by: Jan Zyczkowski <jan.zyczkowski@nordicsemi.no>