[nrf fromlist] drivers: clock_control nrf_lfclk: patch clock option order
The clock options used within the driver are supposed to be ordered
from lowest to highest power consumption, so the lowest/default
option is the most power efficient. The order was reversed to make
the init code of the lfclk a bit simpler, and this was accounted for
in the clock option lookup function. However, the common nrf clock
control request/release feature would request the lowest index, not
the lowest clock option, so the lfclk would default to its highest
power consumption mode.

The clock option init and lookup has been refactored to be sorted
from lowest to highest power consumption, and comments have been
adjusted accordingly.

Upstream PR #: 99382

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>