[nrf fromtree] modules: tfm: don't pad unconfirmed image
If the image is not confirmed, there is no need to pad the entire flash
slot with empty data. This reduces the size of
`tfm_s_zephyr_ns.signed.bin` (the file actually sent for an OTA upgrade)
from 100% of the slot size down to the size of the application.

Signed-off-by: Jordan Yates <jordan@embeint.com>
(cherry picked from commit 6fe6a8bbb473a180299793d4be7d9443f36d441c)