[nrf fromlist] nordic: nrf54h: bicrgen: patch lfosc lfxo accuracy parsing
The accuracy field of lfocs lfxo is incorrectly parsed from the SVD
file. Instead of stripping the "ppm" from the value before converting
it to an int, we are taking only the first three digits. This works
for 500ppm, but not for 20ppm.

Patch to strip last three digits instead.

Upstream PR #: 100506

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>